Fikset en rekke tilbakemeldinger fra PR
