+++ src/@types/index.ts
@@ -1,2 +1,3 @@
 export type { DokumentOversikt, Mottaksretning, FysiskDokument, Dokument, BrevPdfData, SedPdfData } from "./dokument";
 export type { Periode, Lovvalgsperiode, Medlemskapsperiode, OppdaterMedlemskapsperiode } from "./periode";
+export type { Feilmelding } from "./validering";
