+++ src/@types/validering.ts
@@ -1,0 +1,4 @@
+export interface Feilmelding {
+  tittel: string;
+  innhold: string;
+}
