--- src/felleskomponenter/dialogboks/validering/index.js
@@ -1,3 +1,0 @@
-import FeilmeldingDialog from "./feilmeldingDialog";
-
-export default FeilmeldingDialog;
