+++ src/felleskomponenter/stegvelger/validering/Valideringsfeil.test.tsx
@@ -3,15 +3,15 @@ import React from "react";
 
 import * as KV from "../../../kodeverk";
 import MKV from "../../../melosyskodeverk";
-import Valideringsfeil, { ValideringBody } from "./Valideringsfeil";
+import Valideringsfeil, { Feilbeskrivelse } from "./Valideringsfeil";
 
 describe("Valideringsfeil", () => {
   it('Viser feilmelding "Ukjent feil" dersom det ikke finnes en mapping for feilkode', () => {
     const valideringsfeil = shallow(<Valideringsfeil validering={{ kode: "tilfeldigString", felter: [] }} />);
-    const valideringBody = valideringsfeil.find(ValideringBody);
+    const feilbeskrivelse = valideringsfeil.find(Feilbeskrivelse);
 
-    expect(valideringBody).toHaveLength(1);
-    expect(valideringBody.props().tittel).toBe("Ukjent feil");
+    expect(feilbeskrivelse).toHaveLength(1);
+    expect(feilbeskrivelse.props().tittel).toBe("Ukjent feil");
   });
 
   it("viser feilmelding fra kodeverk dersom ingen mapping for feilmelding finnes, og kontrollkode stammer fra manglende utfylling av felter", () => {
@@ -21,11 +21,11 @@ describe("Valideringsfeil", () => {
       />
     );
 
-    const valideringBody = valideringsfeil.find(ValideringBody);
-    expect(valideringBody).toHaveLength(1);
+    const feilbeskrivelse = valideringsfeil.find(Feilbeskrivelse);
+    expect(feilbeskrivelse).toHaveLength(1);
 
-    expect(valideringBody.props().tittel).toBe("Manglende utfylling");
-    expect(valideringBody.props().innhold).toBe(
+    expect(feilbeskrivelse.props().tittel).toBe("Manglende utfylling");
+    expect(feilbeskrivelse.props().innhold).toBe(
       KV.kodeTilTerm(
         MKV.Koder.begrunnelser.kontroll_begrunnelser.MANGLENDE_BOSTEDSADRESSE,
         MKV.KTObjects.begrunnelser.kontroll_begrunnelser
@@ -40,11 +40,11 @@ describe("Valideringsfeil", () => {
       />
     );
 
-    const valideringBody = valideringsfeil.find(ValideringBody);
-    expect(valideringBody).toHaveLength(1);
+    const feilbeskrivelse = valideringsfeil.find(Feilbeskrivelse);
+    expect(feilbeskrivelse).toHaveLength(1);
 
-    expect(valideringBody.props().tittel).toBe("Feil ved kontroll");
-    expect(valideringBody.props().innhold).toBe(
+    expect(feilbeskrivelse.props().tittel).toBe("Feil ved kontroll");
+    expect(feilbeskrivelse.props().innhold).toBe(
       KV.kodeTilTerm(
         MKV.Koder.begrunnelser.kontroll_begrunnelser.INGEN_SLUTTDATO,
         MKV.KTObjects.begrunnelser.kontroll_begrunnelser
