+++ src/felleskomponenter/stegvelger/validering/mapBehandlingsgrunnlagpathTilMenypunkt.test.ts
