+++ src/felleskomponenter/stegvelger/validering/mapBehandlingsgrunnlagpathTilMenypunkt.ts
