+++ src/sider/journalforing/komponenter/feilmeldingDialog.less
@@ -1,4 +1,4 @@
-@import "../../../constants";
+@import "src/constants";
 
 .feilmeldingDialog {
   width: 50%;
